Added pre_save signal to Template class to call "removed_cached_template". Calling this method on pre_save ensures that a previously cached template is deleted from the cache.
Example:
I add a template to override "index.html". I then decide to change the name to "foo.html" to make some changes.
Previously, "dbtemplates::index.html::[site_id]" was not being removed from the cache, and even though the name had changed, the previously cached index.html would be rendered, instead of falling back to the filesystem.

Added try/except in utils.cache "remove_cached_template".
In order to remove a previously cached template, if we change the name of a template, we need to invalidate the previous cache key.
The try/catch is necessary to pass on the case when we are saving a new template.